add login, logout and register to account app
