+++ online_shop/templates/base.html
@@ -47,7 +47,7 @@
                         <a class="nav-link" href="{% url 'store_app:purchases-view' %}">Shopping basket</a>
                     </li>
                     <li class="nav-item">
-                        <a class="nav-link" href="{% url 'store_app:logout-view' %}">Logout</a>
+                        <a class="nav-link" href="{% url 'account:logout-view' %}">Logout</a>
                     </li>
                     <li class="nav-item">
                         <a class="nav-link disabled" href="#">Welcome, {{ user.username }}</a>
@@ -55,10 +55,10 @@
                     {% else %}
 
                     <li class="nav-item">
-                        <a class="nav-link" href="{% url 'store_app:register-view' %}">Register</a>
+                        <a class="nav-link" href="{% url 'account:register-view' %}">Register</a>
                     </li>
                     <li class="nav-item">
-                        <a class="nav-link" href="{% url 'store_app:login-view' %}">Login</a>
+                        <a class="nav-link" href="{% url 'account:login-view' %}">Login</a>
                     </li>
                     {% endif %}
